refactor to fix pr preview build
