+++ web/src/beta/lib/core/Crust/Widgets/Widget/builtin/unsafeWidgets.ts
@@ -1,16 +1,15 @@
 import { merge } from "lodash-es";
 
+import { config } from "@reearth/services/config";
+import { type UnsafeBuiltinPlugin } from "@reearth/services/config/unsafeBuiltinPlugin";
+
 import type { Component } from "..";
-import { UnsafeBuiltinPlugin } from "../../../types";
 
 export type { Component } from "..";
 
 export type UnsafeBuiltinWidgets<T = unknown> = Record<string, T>;
 
-const unsafeBuiltinPlugins = (await import(/* @vite-ignore */ "src/beta/lib/unsafeBuiltinPlugins"))
-  .default;
-
-export const unsafeBuiltinWidgets = processUnsafeBuiltinWidgets(unsafeBuiltinPlugins);
+export const unsafeBuiltinWidgets = processUnsafeBuiltinWidgets(config()?.unsafeBuiltinPlugins);
 
 function processUnsafeBuiltinWidgets(plugin?: UnsafeBuiltinPlugin[]) {
   if (!plugin) return;
